backend/requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
datasets
python-dotenv
faiss
groq
transformers
torch
pypdf
tqdm
fastapi
uvicorn
python-multipart
